// angular16/src/app/model/closejob.model.ts
export interface ICloseJob {
    JobNo: string;
    RealJobNo: string;
    TID: string;
    Bank: string;
    SerialNoEDC: string;
    TechnicName: string;
    RecordDateTime: any;
    UpdateDateTime: any;
    JobType: string;
    JobStatus: string;
    Remark: string;
    PhoneNo: string;
    Merchant: string;
    OldSerialNoEDC: string;
    CustomerName: string;
    CustomerPhoneNo: string;
    InputFileName: string;
    OperationDate: string;
    OperationTime: string;
    OperationEndDate: string;
    OperationEndTime: string;
}